Fixed BusinessPageMobile/converted AddEmpModal into its own file plus mobile version/Fixed manage employee route to ManageEmployeePage instead of Popup
