+++ shiftEase/pages/auth/logOut.js
@@ -1,5 +1,6 @@
 import react from 'react';
 import { useState } from 'react';
+import { useNavigation } from '@react-navigation/native';
 import NavigationManager from '../NavigationManager';
